+++ k8s/helper.go
@@ -3,9 +3,12 @@ package kube
 import (
 	"encoding/json"
 	"fmt"
+	"strings"
 
 	log "github.com/Sirupsen/logrus"
+	"k8s.io/api/core/v1"
 	"k8s.io/api/extensions/v1beta1"
+	"k8s.io/apimachinery/pkg/api/resource"
 )
 
 func ingressNamespace(parentNS string) string {
@@ -70,4 +73,40 @@ func mergeIngressRules(dest *v1beta1.Ingress, source *v1beta1.Ingress) *v1beta1.
 	log.Debugf("final ingress rules %s", toJson(dest.Spec.Rules))
 
 	return dest
+}
+
+// mergeResourceConstraints sets the limit and requests values for cpu, memory for a container
+func mergeResourceConstraints(resourceType v1.ResourceName, container *v1.Container, reqLimit string) error {
+	resourceLimit, resourceRequest := container.Resources.Limits, container.Resources.Requests
+
+	if reqLimit != "" {
+		parts := strings.Split(reqLimit, "/")
+		var request, limit string
+		if len(parts) == 2 {
+			request = parts[0]
+			limit = parts[1]
+		} else {
+			limit = parts[0]
+		}
+
+		if request != "" {
+			value, err := resource.ParseQuantity(request)
+			if err != nil {
+				return err
+			}
+
+			resourceRequest[resourceType] = value
+		}
+
+		if limit != "" {
+			value, err := resource.ParseQuantity(limit)
+			if err != nil {
+				return err
+			}
+
+			resourceLimit[resourceType] = value
+		}
+	}
+
+	return nil
 }