+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - [CLI] Check build status before deploying an app
 ### Added
 - [CLI] `datacol env unset` supports multiple parameters
+- Default memory limits (256Mi/512Mi) for any container
 
 ## [1.0.0-alpha.13] - 2018-04-21
 ### Fixed